+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### 20210715
+
+* Updated documentation
+* Fixed compresison level for LZMA compressor
+
+### 20210714
+
+* Added lzma compression to data and audio streams
+
 ### 20210707
 
 * Added zlib compression to audio stream